base for parameters parsing

+++ Command/Scripts/Parameters/NumberParameter.cs
@@ -3,12 +3,12 @@ using Command.Parameters.ParameterParsers;
 
 namespace Command.Parameters
 {
-    public class NumberParameter : Parameter<byte>
+    public class NumberParameter : Parameter<short>
     {
-        protected override IParameterParser<byte> Parser { get; set; } = new NumberParser();
+        protected override IParameterParser<short> Parser { get; set; } = new NumberParser();
 
 
-        public static explicit operator NumberParameter(byte number)
+        public static explicit operator NumberParameter(short number)
         {
             return new NumberParameter().GetParameter(number.ToString()) as NumberParameter;
         }
